requirements.txt
# Core ML
torch>=2.0.0
torchaudio>=2.0.0
torchtext>=0.15.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
deepfilternet>=0.5.0
ffmpeg-normalize>=1.27.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Text Processing
phonemizer>=3.2.0

# Visualization & Logging
tensorboard>=2.13.0
matplotlib>=3.7.0

# Hugging Face
huggingface_hub>=0.19.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Development
ipykernel>=6.0.0
jupyter>=1.0.0
